[WIP] Add pie charts for percentages

+++ app/src/main/java/me/benguiman/spainstats/domain/models/DataSeries.kt
@@ -48,7 +48,7 @@ object AveragePopulationAgeSeries : DataSeries {
         BaseData
     )
     override val headlineVariable = AveragePopulationAge
-    override val dataType = DataType.DOUBLE
+    override val dataType = DataType.INTEGER
 }
 
 object IpvaAnnualVariation : DataSeries {
